app/requirements.txt
fastapi 
uvicorn[standard]
python-dotenv
confluent-kafka
fastavro
pydantic
pydantic-settings
airbyte-api==0.50.0